Update README.md
In this commit, I've updated my GitHub README to offer a more comprehensive view of my background, interests, and ongoing projects such as Introduction, Languages and Tools, Connect with Me, Fun Fact and GitHub Stats.
I look forward to connecting and collaborating with fellow developers and enthusiasts on this platform. Feel free to reach out and join me in my coding and learning journey. 🚀